UX-59 new icons for Reddit, QQ, Tencent Weibo
Traced vector icons for these 3 additional providers. The QQ one is
kinda weird (the outline of the penguin), but I found it used in a few
places where they had a monochrome logo. The alternate version where the
penguin shape is solid was typically used where the icon was reversed
out of a different shape, which didn't apply as well here.